+++ server/netty/src/com/riiablo/net/reliable/channel/ReliableMessageChannel.java
@@ -22,7 +22,7 @@ public class ReliableMessageChannel extends MessageChannel {
   }
 
   @Override
-  public void update(long delta, DatagramChannel ch) {
+  public void update(float delta, DatagramChannel ch) {
 
   }
